+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 The format is based on and uses the types of changes according to [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [vNext]
+
+### Added (v5.0)
+
+- SPInstallPrereqs
+  - Added support for SharePoint vNext
+
 ## [Unreleased]
 
 ## [4.8.0] - 2021-08-31